+++ packages/terra-list/src/ListSubsectionHeader.jsx
@@ -107,7 +107,8 @@ const ListSubsectionHeader = ({
     attrSpread.onBlur = ListUtils.wrappedEventCallback(onBlur, event => event.currentTarget.setAttribute('data-item-show-focus', 'true'));
     attrSpread.onMouseDown = ListUtils.wrappedEventCallback(onMouseDown, event => event.currentTarget.setAttribute('data-item-show-focus', 'false'));
   }
-
+  /* eslint-disable-next-line no-param-reassign */
+  delete customProps?.isTabFocusDisabled;
   return (
     <li {...customProps} className={cx('list-item', theme.className)}>
       <Element className={cx('title')}>