+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
+* Changed
+  * Removed `ariaDescription` example added for terra-list
+
 * Added
   * Added a test case for `terra-list` to support drag and drop in `terra-modal-manager`.
-  * Added aria-label to announce the visual label.
+  * Added a invalid example in `terra-search-Field`.
+  * Added aria-label to announce the visual label for `terra-form-select`.
 
 ## 1.52.0 - (November 21, 2023)